+++ modules/framegrabber/include/ifm3d/fg/frame.h
@@ -179,6 +179,11 @@ namespace ifm3d
     */
     O3R_ODS_POLAR_OCC_GRID = static_cast<uint64_t>(ifm3d::image_chunk::O3R_ODS_POLAR_OCC_GRID),
 
+    /** @hideinitializer
+     * @brief ifm3d::ODSExtrinsicCalibrationCorrectionV1
+    */
+    O3R_ODS_EXTRINSIC_CALIBRATION_CORRECTION = static_cast<uint64_t>(ifm3d::image_chunk::O3R_ODS_EXTRINSIC_CALIBRATION_CORRECTION),
+
     /** @hideinitializer 
      * @brief The point cloud encoded as a 3 channel XYZ image
     */