+++ ChangeLog.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - l4t based .deb packages
 
+### Added
+- Add struct `ODSExtrinsicCalibrationCorrectionV1` in Deserialize module
+
 ## 1.6.7 - 2025-03-28
 - Update the CLI documentation